Added explicit sizes

+++ src/data/data.hpp
@@ -121,8 +121,8 @@ struct EmergencyBrakes : public Module {
 
 struct Motors : public Module {
   static constexpr int kNumMotors          = 4;
-  static constexpr int kMaximumTemperature = 150;
-  static constexpr int kMaximumCurrent     = 1500;  // mA
+  static constexpr uint8_t kMaximumTemperature = 150;
+  static constexpr uint16_t kMaximumCurrent     = 1500;  // mA
   std::array<uint32_t, kNumMotors> rpms    = {{0, 0, 0, 0}};
 };
